split account view component from table and share with recieve view
