+++ src/renderer/components/dashboard/accounts.tsx
@@ -1,15 +1,14 @@
 import React from 'react';
-import CopyToClipboard from 'react-copy-to-clipboard';
-import ReactTooltip from 'react-tooltip';
 import * as zfaceHelper from '../../zface_helper';
-import AccountSelect from './account_select';
+import AccountSelect from './../accounts/account_select';
+import AccountTable from './../accounts/account_table';
 
 interface IAddressStableStates {
     accounts: zfaceHelper.AccountInfo[];
     defaultAccountName: string;
 }
 
-export default class AddressTable extends React.Component<{}, IAddressStableStates> {
+export default class Accounts extends React.Component<{}, IAddressStableStates> {
     public constructor(props: any) {
         super(props);
         this.state = {
@@ -21,7 +20,6 @@ export default class AddressTable extends React.Component<{}, IAddressStableStat
     public render() {
         return (
             <div>
-                <h2>Your Accounts</h2>
                 <div className="row">
                     <div className="col-sm-6">
                         <AccountSelect
@@ -31,32 +29,7 @@ export default class AddressTable extends React.Component<{}, IAddressStableStat
                         />
                     </div>
                 </div>
-                <table className="table">
-                    <tbody>
-                        <tr>
-                            <th>Name</th>
-                            <th>Address</th>
-                            <th></th>
-                        </tr>
-                        {this.state.accounts.map((account) =>
-                            <tr key={account.name}>
-                                <td>
-                                    { account.isDefault && <div className="font-weight-bold">*{account.name}</div> }
-                                    { !account.isDefault && <div>{account.name}</div> }
-                                </td>
-                                <td>{account.address}</td>
-                                <td>
-                                    <CopyToClipboard text={account.address}>
-                                        <a data-tip="click to copy">
-                                            <i className="fas fa-copy fa-lg copy-icon"></i>
-                                        </a>
-                                    </CopyToClipboard>
-                                    <ReactTooltip />
-                                </td>
-                            </tr>,
-                        )}
-                    </tbody>
-                </table>
+                <AccountTable accounts={this.state.accounts} />
             </div>
         );
     }
